Calculate Y cordinate offset based on BGRT configuration.
This commit improves the position of the vendor image, making it
dependent on the Y offset took from BGRT configuration. From now,
the image should be placed in the same Y position as BIOS logo,
making it to look much better on transition between BIOS POST logo
and bootsplash. As X position is basically calculated to make image
centered, it may even match the exact same position of BIOS logo,
which gives the user a possibility of configuring his boot enviroment
to make a perfectly smooth transition between splashes of different
parts of the boot process.